comment entity finished, edit/delete per author or admin implemented and validation work

+++ gw-nestjs/src/posts/posts.service.ts
@@ -3,6 +3,9 @@ import {
   BadRequestException,
   InternalServerErrorException,
   NotFoundException,
+  Inject,
+  Scope,
+  ForbiddenException,
 } from '@nestjs/common';
 import { Post } from './models/post.entity';
 import { Repository, DeepPartial } from 'typeorm';
@@ -12,16 +15,23 @@ import { ResponseError } from 'src/common/models/response';
 import { QueryRequest } from 'src/common/models/query-request';
 import { PagedData } from 'src/common/models/paged-data';
 import { GetPostDto } from './models/dto/get-post.dto';
+import { REQUEST } from '@nestjs/core';
+import { Request } from 'express';
+import { TokenUserPayloadDto } from 'src/auth/models/dto/token-user-payload.dto';
+import { UserRole } from 'src/users/models/user.entity';
 
 @Injectable()
 export class PostsService {
   constructor(
     @InjectRepository(Post) private postsRepository: Repository<Post>,
     private usersService: UsersService,
+    @Inject(REQUEST) private request: Request,
   ) {}
 
   async create(payload: Post): Promise<Post> {
     try {
+      const user = new TokenUserPayloadDto(this.request.user);
+      payload.author = await this.usersService.findOne({ id: user.id });
       return await this.postsRepository.save(payload);
     } catch (error) {
       throw new BadRequestException(
@@ -59,8 +69,9 @@ export class PostsService {
     return post;
   }
 
-  async update(id: string, model: Post) {
+  async update(id: string, model: Post): Promise<Post> {
     const post = await this.findOne({ id });
+    this.authorize(post);
     post.content = model.content;
     post.title = model.title;
     post.isPublished = model.isPublished;
@@ -73,10 +84,21 @@ export class PostsService {
 
   async delete(params: DeepPartial<Post>): Promise<Post> {
     const post = await this.findOne(params);
+    this.authorize(post);
     try {
       return await this.postsRepository.remove(post);
     } catch (error) {
       throw new InternalServerErrorException(error.toString());
+    }
+  }
+
+  private authorize(entity: Post) {
+    const tokenUser = new TokenUserPayloadDto(this.request.user);
+    if (
+      tokenUser.id !== entity.author.id &&
+      tokenUser.role !== UserRole.ADMIN
+    ) {
+      throw new ForbiddenException();
     }
   }
 }
